style: view control 파일들 이름 변경
